+++ algs/load_grid_output.py
@@ -3,6 +3,7 @@ import os
 import glob
 import re
 import json
+import time
 
 from qgis.PyQt.QtWidgets import (
     QWizard,
@@ -435,21 +436,12 @@ class ScenarioSelectionPage(QWizardPage):
         """
         scenarios = set()
         events = set()
+        # Keywords to ignore when extracting scenarios, as they are commonly present in filenames but not scenario identifiers
         ignored_keywords = {
-            "h",
-            "d",
-            "v",
-            "q",
-            "hr",
-            "max",
-            "tmax",
-            "min",
-            "avg",
-            "dem",
-            "grid",
-            "znz2",
-            "tmax_h",
-            "zaem1",
+            "h","d","v","q",
+            "hr","max","tmax","min",
+            "avg","dem","grid","tmax_h",
+            "znz2","zaem1","zud1"
         }
 
         parts = filename_str.split("_")
@@ -815,13 +807,17 @@ class PreviewPage(QWizardPage):
 
     def init_ui(self):
         layout = QVBoxLayout()
-        self.preview_table = QTableWidget(0, 2)
-        self.preview_table.setHorizontalHeaderLabels(["Select", "Grid File"])
+        self.preview_table = QTableWidget(0, 3)
+        self.preview_table.setHorizontalHeaderLabels(["Select", "Grid File", "Modified Date"])
         self.preview_table.horizontalHeader().setSectionResizeMode(0, QHeaderView.Fixed)
         self.preview_table.setColumnWidth(0, 60)
         self.preview_table.horizontalHeader().setSectionResizeMode(
             1, QHeaderView.Stretch
         )
+        self.preview_table.horizontalHeader().setSectionResizeMode(
+            2, QHeaderView.ResizeToContents
+        )
+        self.preview_table.setSortingEnabled(True)
         layout.addWidget(self.preview_table)
 
         self.empty_warning_label = QLabel(
@@ -901,20 +897,30 @@ class PreviewPage(QWizardPage):
                     _write_debug_log("\n".join(debug_lines))
 
                 self.grid_files_to_load = sorted(unique_paths)
+                self.preview_table.setSortingEnabled(False)
                 self.preview_table.setRowCount(len(self.grid_files_to_load))
 
                 for i, fpath in enumerate(self.grid_files_to_load):
                     cb = QCheckBox()
                     cb.setChecked(True)
                     self.preview_table.setCellWidget(i, 0, cb)
+                    
                     item = QTableWidgetItem(os.path.basename(fpath))
                     item.setData(Qt.UserRole, fpath)
+                    item.setFlags(item.flags() & ~Qt.ItemIsEditable)
                     self.preview_table.setItem(i, 1, item)
+                    
+                    mtime_str = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(os.path.getmtime(fpath)))
+                    date_item = QTableWidgetItem(mtime_str)
+                    date_item.setFlags(date_item.flags() & ~Qt.ItemIsEditable)
+                    self.preview_table.setItem(i, 2, date_item)
 
                 if len(self.grid_files_to_load) == 0:
                     self.empty_warning_label.setVisible(True)
                 else:
                     self.empty_warning_label.setVisible(False)
+                    
+                self.preview_table.setSortingEnabled(True)
 
         finally:
             self.preview_table.setUpdatesEnabled(True)